+++ crates/ibc/src/core/mod.rs
@@ -27,6 +27,7 @@ pub mod ics03_connection;
 pub mod ics04_channel;
 pub mod ics23_commitment;
 pub mod ics24_host;
+pub mod router;
 
 pub mod events;
 pub mod timestamp;
@@ -34,14 +35,12 @@ pub mod timestamp;
 mod context;
 mod handler;
 mod msgs;
-pub(crate) mod router;
 
 pub use handler::dispatch;
 pub use handler::execute;
 pub use handler::validate;
 
 pub use context::*;
-pub use router::*;
 
 pub use msgs::Msg;
 pub use msgs::MsgEnvelope;
